go.net/html/charset: encoding names
Lookup now returns the canonical name as well as the Encoding.

This will make it easier for users to discover what encoding they
actually have as a return value from functions in this package.
They will also be able to store the name for re-use.

R=nigeltao, mpvl
CC=golang-dev
https://codereview.appspot.com/30090043

Committer: Nigel Tao <nigeltao@golang.org>